[project @ b8ccf9e5e42d9d9de1a03bf60eeba66ff3e966f4] Updat copyright notices in stable.
Original author: Thomas Roessler <roessler@does-not-exist.org>
Date: 2000-01-06 11:58:16+00:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@1150 2bb92213-1a62-4c20-bdc6-606bde718ed5

+++ pattern.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1996-9 Michael R. Elkins <me@cs.hmc.edu>, and others
+ * Copyright (C) 1996-2000 Michael R. Elkins <me@cs.hmc.edu>, and others
  * 
  *     This program is free software; you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
@@ -143,7 +143,7 @@ msg_search (CONTEXT *ctx, regex_t *rx, char *buf, size_t blen, int op, int msgno
       /* decode the header / body */
       memset (&s, 0, sizeof (s));
       s.fpin = msg->fp;
-      s.flags = M_CHARCONV;
+      s.flags |= M_CHARCONV;
       mutt_mktemp (tempfile);
       if ((s.fpout = safe_fopen (tempfile, "w+")) == NULL)
       {
@@ -164,8 +164,11 @@ msg_search (CONTEXT *ctx, regex_t *rx, char *buf, size_t blen, int op, int msgno
 	if (h->pgp & PGPENCRYPT && !pgp_valid_passphrase())
 	{
 	  mx_close_message (&msg);
-	  fclose (fp);
-	  unlink (tempfile);
+	  if (fp)
+	  {
+	    fclose (fp);
+	    unlink (tempfile);
+	  }
 	  return (0);
 	}
 #endif
@@ -765,10 +768,10 @@ static int match_reference (regex_t *rx, LIST *refs)
 int mutt_is_list_recipient (int alladdr, ADDRESS *a1, ADDRESS *a2)
 {
   for (; a1 ; a1 = a1->next)
-    if (alladdr ^ mutt_is_subscribed_list (a1))
+    if (alladdr ^ mutt_is_mail_list (a1))
       return (! alladdr);
   for (; a2 ; a2 = a2->next)
-    if (alladdr ^ mutt_is_subscribed_list (a2))
+    if (alladdr ^ mutt_is_mail_list (a2))
       return (! alladdr);
   return alladdr;
 }
